Move password handler to initialize step
