+++ controllers/api/v1/token.go
@@ -2,16 +2,12 @@ package apiv1controllers
 
 import (
 	"encoding/json"
-	"fmt"
 	"net/http"
 	"net/http/httptest"
 
 	. "go-crawler-challenge/controllers/api"
-	form "go-crawler-challenge/forms/session"
 	v1serializers "go-crawler-challenge/serializers/v1"
 	"go-crawler-challenge/services/oauth"
-
-	"github.com/go-oauth2/oauth2/v4/errors"
 )
 
 // TokenController operations for Token
@@ -30,8 +26,6 @@ func (c *TokenController) URLMapping() {
 // @Success 200
 // @router /api/v1/oauth/token [post]
 func (c *TokenController) Create() {
-	oauth.ServerOauth.SetPasswordAuthorizationHandler(passwordAuthorizationHandler)
-
 	writer := httptest.NewRecorder()
 
 	err := oauth.ServerOauth.HandleTokenRequest(writer, c.Ctx.Request)
@@ -71,18 +65,4 @@ func (c *TokenController) handleResponseError(writer *httptest.ResponseRecorder)
 	} else {
 		c.RenderError(http.StatusText(writer.Code), tokenErrorObject.Detail, writer.Code, tokenErrorObject.Code)
 	}
-}
-
-func passwordAuthorizationHandler(email string, password string) (string, error) {
-	authenticationForm := form.AuthenticationForm{
-		Email:    email,
-		Password: password,
-	}
-
-	user, errorList := authenticationForm.Authenticate()
-	if len(errorList) > 0 {
-		return "", errors.ErrInvalidClient
-	}
-
-	return fmt.Sprint(user.Id), nil
 }
